requirements.txt
asgiref==3.5.2
boto3==1.24.53
botocore==1.27.53
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.0.0
dj-rest-auth==2.2.5
Django==4.1
django-allauth==0.51.0
django-ckeditor==6.5.0
django-coreapi==2.3.0
django-cors-headers==3.13.0
django-filter==22.1
django-heroku==0.3.1
django-import-export==2.8.0
django-js-asset==2.0.0
django-postgresql==0.0.3
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupPy==1.14
MarkupSafe==2.1.1
oauthlib==3.2.1
odfpy==1.4.1
openpyxl==3.0.10
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
stripe==4.1.0
tablib==3.2.1
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.11
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
